Don't depend on ppx_cstruct, add mli file
The cstruct ppx was expanded and manually added in the source file. It
was only used in one place for a relatively small struct. The hexdump
function was omitted.

A lib/rawlink_lowlevel.mli file was added with all members exposed. The
intent is to eventually unexpose what is not needed.